Remove ManufacturerLock input from edit form on clone page
This was only present on the export form, but it would always match the profile being cloned - meaning that if you clone a manufacturer locked profile in export mode the new profile would also be locked. Basic mode this issue wasn't present because the form didn't set that value in a hidden field.